Masters_project/authentication/python/requirements.txt
cryptography==44.0.2
mysql-connector-python==9.2.0